+++ packages/ui/package.json
@@ -31,6 +31,7 @@
     "dequal": "^2.0.3",
     "framer-motion": "^11.0.8",
     "medium-zoom": "^1.1.0",
+    "p-map": "^5.5.0",
     "property-information": "^6.4.0",
     "radash": "^10.7.1",
     "react-hook-form": "^7.51.0",
